% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_content.R
\docType{data}
\name{content}
\alias{content}
\title{Data content}
\format{An object of class \code{list} of length 1.}
\usage{
content
}
\description{
A JSON format list of OPS biblio data with 94 entries. For testing with ops_biblio().
}
\keyword{datasets}

