% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_countrycode_data.R
\docType{data}
\name{countrycode_data}
\alias{countrycode_data}
\title{Data countrycode_data}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 279 rows and 17 columns.}
\usage{
countrycode_data
}
\description{
A data file containing the table used in ops_country()
}
\keyword{datasets}

