% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_epodoc.R
\docType{data}
\name{epodoc}
\alias{epodoc}
\title{Data epodoc}
\format{An object of class \code{response} of length 10.}
\usage{
epodoc
}
\description{
A JSON file containing 94 raw bibliographic records from OPS. Used for for testing with ops_biblio.
}
\keyword{datasets}

