% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops_auth.R
\name{ops_auth}
\alias{ops_auth}
\title{Authenticate with Open Patent Services (OPS).}
\usage{
ops_auth(key, secret)
}
\arguments{
\item{key}{Alphanumeric provided by OPS (quoted).}

\item{secret}{Alphanumeric secret provided by OPS (quoted).}
}
\value{
200 if authenticated, 400 if not. See error messages for information.
}
\description{
The European Patent Office Open Patent Services (OPS) requires
authentication. First, register for a free account at
\url{http://www.epo.org/searching-for-patents/technical/espacenet/ops.html#tab1}.
Then obtain a key and secret by creating an App under My Apps in the top
right of the API Console for your account. Then copy the key and secret and either paste into ops_auth (quoted) or store in the environment as key and secret.
}
\examples{
\dontrun{ops_auth(key, secret)}
}

