% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops_biblio.R
\name{ops_biblio}
\alias{ops_biblio}
\title{Process Multiple Lists of OPS Bibliographic Data (formerly ops_multi_biblio)}
\usage{
ops_biblio(x)
}
\arguments{
\item{x}{A list of lists containing bibliographic data from OPS}
}
\value{
a data.frame
}
\description{
Use this function to process lists containing multiple lists of
  bibliographic data from OPS into a single data frame. Typically, the object
  will contain the results of multiple calls to OPS for bibliographic data in
  chunks of 100 records (see ops_urls()). For example, 10 calls will generate
  a single list containing 10 list objects each containing 100 records.
}
\details{
The function uses rapply and later aggregate. Note that aggregate
  will generate a warning. The use of fill() from tidyr is a temporary
  measure and is a likely source of weakness in the code. Note that in order
  to produce data.frames with an equal number of columns across all years the
  `ipc` column for the historic IPC7 classification is dropped where present.
}
\examples{
\dontrun{ops_biblio(ops_biblios)}
}

