% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops_biblio_rename.R
\name{ops_biblio_rename}
\alias{ops_biblio_rename}
\title{Convert OPS biblio names in Rows from JSON format to sane names (internal function)}
\usage{
ops_biblio_rename(x)
}
\arguments{
\item{x}{a dataframe with row names to be converted from JSON to sane text.}
}
\value{
a data.frame with column 1 rows renamed.
}
\description{
Used internally. Assumes a data.frame with two columns. Column 1 contains JSON strings as rows to be renamed. Column 2 contains the data. The function replaces Column 1 row content with the dictionary entry in `ops_dictionary` using dplyr::left_join(). To convert column names (rather than row names) use ops_rename(). Note check consistency between ops_dictionary and ops_rename.
}
\details{
calls ops_dictionary to convert OPS JSON strings to normal names. ops_dictionary must be available for this to work.
}
\examples{
\dontrun{ops_biblio_rename(df)}
}

