% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ops_biblios.R
\docType{data}
\name{ops_biblios}
\alias{ops_biblios}
\title{Data pizza bibliographic data (list)}
\format{An object of class \code{list} of length 9.}
\source{
EPO Open Patent Services through ops_urls ((urls_1990_2000 <-
  ops_urls("pizza", type = "ta", start = 1990, end = 2000))
}
\usage{
ops_biblios
}
\description{
A JSON list from a call to the OPS bibliographic (biblio) service using ops_iterate. For use with ops_biblio()
}
\keyword{datasets}

