% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops_country.R
\name{ops_country}
\alias{ops_country}
\title{Convert two letter patent country codes to country names}
\usage{
ops_country(data, col)
}
\arguments{
\item{data}{A data.frame containing a country column to}

\item{col}{a column named "country" containing two letter country codes}
}
\value{
A data.frame
}
\description{
This function converts two letter country and patent office
  codes to country and office names. It converts the column and adds a
  `country_name` column with the full country name.
}
\examples{
\dontrun{ops_country(data = pizza_numbers, col = "country")}
}
\references{
The base country list is imported from the countrycode package
  (on CRAN) in the countrycode_data.csv file
  \url{https://github.com/vincentarelbundock/countrycode/raw/master/data/countrycode_data.csv}.
  That file is reproduced and modified with entries in the iso2c field for
  regional patent offices from WIPO standard St.3
  \url{http://www.wipo.int/pct/guide/en/gdvol1/annexes/annexk/ax_k.pdf}.
  Continent and region fields are assigned to patent at present.
}

