% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ops_dictionary.R
\docType{data}
\name{ops_dictionary}
\alias{ops_dictionary}
\title{Data ops_dictionary}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 126 rows and 2 columns.}
\usage{
ops_dictionary
}
\description{
A data file containing the dictionary used on ops_dictionary for converting JSON strings to sensible names.
}
\keyword{datasets}

