% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops_fetch_biblio.R
\name{ops_fetch_biblio}
\alias{ops_fetch_biblio}
\title{Search and retrieve bibliographic patent data from OPS}
\usage{
ops_fetch_biblio(query, type, service, start, end, timer)
}
\arguments{
\item{query}{A search query (quoted).}

\item{type}{Select from ti (title), ta (title and abstract), quoted.}

\item{service}{Limited to "biblio"}

\item{start}{The year or date to start with.}

\item{end}{The year or date to end with.}

\item{timer}{The number of seconds between each call for 100 results (numeric)}
}
\value{
A data frame
}
\description{
Use to search and retrieve bibliographic data from OPS and create a data.frame with the results. Use ops_count to establish the total number of results and year limits. Will not presently retrieve over 2000 results.
}
\details{
A warning may be generated that "There were 50 or more warnings...". This is expected behaviour and arises from the use of aggregate() in ops_biblio inside this function.
}
\examples{
\dontrun{drones <- ops_fetch_biblio(query = "drones", type = "ta", service = "biblio", start = 1990, end = 2015, timer = 10)}
}

