% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops_filter.R
\name{ops_filter}
\alias{ops_filter}
\title{Filter data frame for countries in OPS with Full Text availability}
\usage{
ops_filter(data)
}
\arguments{
\item{data}{A data.frame containing a column "country" or "publication_country" with two letter country and patent office codes (e.g. EP, WO).}
}
\value{
a data.frame
}
\description{
OPS presently includes patent full texts for documents published by Austria (AT), Canada (CA), the European Patent Office (EP), Great Britain (GB), the Patent Cooperation Treaty (WO) and Spain (ES). These are the only countries with full text. Queries to ops_fulltext() that include other countries will fail. This function takes a list of country codes or patent numbers containing country codes and filters them for jurisdictions with full text access.
}
\examples{
\dontrun{ops_filter(data)}
}

