% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops_iterate.R
\name{ops_iterate}
\alias{ops_iterate}
\title{iterate - internal (this works for WHAT?)}
\usage{
ops_iterate(data, service = "", type = NULL, timer = 20)
}
\arguments{
\item{data}{A query consisting of either urls for use with the numbers service or patent numbers for use with the fulltext service. See details.}

\item{service}{Character, select from "numbers"(to retrieve numbers from a text query) or "fulltext" to use a vector of numbers.}

\item{type}{The type of search for "fulltext" only. either "fulltext", "description" or "claims".}

\item{timer}{The delay before sending each request}
}
\value{
a JSON response object for use in other functions
}
\description{
Used internally to send calls to OPS
}
\examples{
\dontrun{threepubs <- c("WO0000034", "WO0000035", "WO0005967")
ft <- iterate(threepubs, service = "fulltext", type = "fulltext")
ft <- iterate(threepubs, type = "claims")}
}

