% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops_numbers.R
\name{ops_numbers}
\alias{ops_numbers}
\title{Create a data frame of results from JSON numbers Open Patent Services}
\usage{
ops_numbers(data, espacenet = FALSE, lens = FALSE, patentscope = FALSE,
  google = FALSE)
}
\arguments{
\item{data}{A JSON formatted response object from the OPS Numbers service.}

\item{espacenet}{Add a link to the espacenet database Default is TRUE.}

\item{lens}{Add a link to the lens database. Default is TRUE.}

\item{patentscope}{Add a link to the WIPO patentscope database. Default is TRUE.}

\item{google}{Add a link to the google_patents database. Default is TRUE.}
}
\value{
A dataframe of results from Open Patent Services.
}
\description{
A JSON formatted response from the European Patent Office Open
  Patent Services (OPS)
}
\details{
Arguments are provided to create hyperlinks from a variety of free
  patent databases. Note that coverage varies between databases and, with the
  exception of espacenet, it should not be expected that all hyperlinks will
  lead to the documents. The epodoc format as used here is effectively the
  straight application number minus the kind code.
}
\examples{
\dontrun{ops_numbers(pizza) #type response}
}

