% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops_numbers_.R
\name{ops_numbers_}
\alias{ops_numbers_}
\title{(internal) Process list from ops_publications to data.frame}
\usage{
ops_numbers_(content)
}
\arguments{
\item{content}{A list generated from a call to OPS using ops_publications.}
}
\value{
a data frame
}
\description{
Process a list from a call to ops_publications (the numbers service) into a data.frame. Used internally in ops_publications.
}
\examples{
\dontrun{ops_numbers(pizza)}
}

