% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops_post.R
\name{ops_post}
\alias{ops_post}
\title{Retrieve bibliographic information using patent numbers}
\usage{
ops_post(x)
}
\arguments{
\item{x}{a character vector containing patent numbers produced by
ops_numbers(), epodoc column preferred.}
}
\value{
a JSON format response
}
\description{
Use this function to fetch patent front pages 'biblios' from OPS
  by posting a list of publication numbers using httr::POST.
}
\details{
epodoc format appears to capture more data than docdb but
  experimentation may be needed.
}
\examples{
\dontrun{x <- more_pizza$epodoc}
}

