% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops_publications.R
\name{ops_publications}
\alias{ops_publications}
\title{Search and retrieve patent numbers}
\usage{
ops_publications(query = "", type = "NULL", start = NULL, end = NULL)
}
\arguments{
\item{query}{A CQL query.}

\item{type}{character, ti (title), ta (title & abstract), biblio (default).}

\item{start}{- YYYY or YYYYMMDD, publication date.}

\item{end}{- YYYY or YYYYMMDD, publication date.}
}
\value{
A data.frame of patent numbers.
}
\description{
Use a text query to retrieve patent numbers for use in other opsr functions such as retrieving biblios, family data, full texts, register or legal status information.
}
\details{
The basic workflow for OPS is to perform a text query, retrieve patent numbers and then gather other information (biblios, family data etc.). This function combines ops_urls, ops_count, ops_iterate, ops_numbers and ops_country in one function.
}
\examples{
\dontrun{more_pizza <- ops_publications(query = "pizza", type = "ta", start = 1990, end = 2000)}
}

