% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops_status.R
\name{ops_status}
\alias{ops_status}
\title{Check OPS Authentication and Throttle Status}
\usage{
ops_status(x = NULL)
}
\arguments{
\item{x}{unquoted character x to run the function.}
}
\value{
numeric, 200 = all good, other = problem. For the throttle green is good, scores lower than 100 mean a service is busy.
}
\description{
Search query to test status of response. Loook for 200. If other rerun
authentication or use ops_throttle for detailed status. If 403 check fair
usage in headers. Throttle data indicates the status of a particular service at a particular time. If a query fails it may be because a service is under heavy load.
}
\examples{
ops_status()
}

