% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops_urls.R
\name{ops_urls}
\alias{ops_urls}
\title{Generate URLs to search OPS biblios}
\usage{
ops_urls(query = "", type = "NULL", start = NULL, end = NULL)
}
\arguments{
\item{query}{- character}

\item{type}{- character, ti (title), ta (title & abstract), biblio (default)}

\item{start}{- YYYY or YYYYMMDD, publication date.}

\item{end}{- YYYY or YYYYMMDD, publication date}
}
\value{
prints the total number of results for a query and creates a character vector with urls for that range.
}
\description{
This function generates a set of URLs with a maximum of 100 results per URL for searching either OPS titles, titles & abstracts, biblios (default) and/or date ranges (publication date). See Details.
}
\details{
The OPS service permits a maximum of 100 records per URL and an absolute hard maximum of 2000 results per set of URLs. This means that to retrieve a complete set of data for a query a set of URLs must be generated calling upto 100 results in sequence (1-100, 101-200 etc.) upto the maximum of 2000 per query. The query will fail if more than 2000 results are included. ops_urls generates the urls and splits them into a list divided into groups of urls with ranges under 2000. That list can then be used with ops_iterate (to fetch the patent numbers) and ops_parse or run as one with ops_publications.

ops_urls() will print the number of results for a given query and year range and create a vector of URLS. Where there are large numbers of results (over 2000) start by using ops_count() to work out the total number of results for a query and then use it to work out year ranges under 2000 results for input into the start and end arguments of ops_urls().
}
\examples{
\dontrun{urls <- ops_urls(query = "pizza", type = "ti", start = 1990, end = 2000)}
}

