% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_numbers_results_.R
\name{parse_numbers_results_}
\alias{parse_numbers_results_}
\title{internal Parse return from the numbers service from melt into data.frame}
\usage{
parse_numbers_results_(results)
}
\arguments{
\item{results}{a data.frame}
}
\value{
a data frame
}
\description{
This is used internally to convert the results of a call to ops_publications containing more than 100 results into a data.frame. The input in ops_publications is a data.frame from reshape2.
}
\examples{
\dontrun{parse_numbers_results(results)}
}

