% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_pizza.R
\docType{data}
\name{pizza}
\alias{pizza}
\title{Data pizza}
\format{An object of class \code{list} of length 1.}
\usage{
pizza
}
\description{
A JSON format list of OPS biblio data. For testing with ops_biblio().
}
\keyword{datasets}

