% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_pizza2.R
\docType{data}
\name{pizza2}
\alias{pizza2}
\title{Data pizza2}
\format{An object of class \code{list} of length 15.}
\usage{
pizza2
}
\description{
Second part of JSON format list of OPS biblio data. For testing with ops_biblio().
}
\keyword{datasets}

