% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_pizza3.R
\docType{data}
\name{pizza3}
\alias{pizza3}
\title{Data pizza3}
\format{An object of class \code{list} of length 7.}
\usage{
pizza3
}
\description{
Third part of JSON format list of OPS biblio data. For testing with ops_biblio().
}
\keyword{datasets}

