% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_pizza_biblios.R
\docType{data}
\name{pizza_biblios}
\alias{pizza_biblios}
\title{Data pizza bibliographic (biblio)}
\format{An object of class \code{list} of length 9.}
\source{
EPO Open Patent Services through ops_iterate(urls_1990_2000, service = "biblio", timer = 30)
}
\usage{
pizza_biblios
}
\description{
A sample of OPS bibliographic (front page data)
}
\keyword{datasets}

