% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_pizza_complete.R
\docType{data}
\name{pizza_complete}
\alias{pizza_complete}
\title{Data pizza_complete}
\format{An object of class \code{data.frame} with 2981 rows and 69 columns.}
\usage{
pizza_complete
}
\description{
combined pizza dataset as single data.frame from pizza, pizza2, pizza3 processed with ops_biblio()
}
\keyword{datasets}

