% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_pub.R
\docType{data}
\name{pub}
\alias{pub}
\title{Data pub}
\format{An object of class \code{character} of length 1.}
\usage{
pub
}
\description{
A single WO epodoc patent document number for use in testing.
}
\keyword{datasets}

