% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_publication_numbers.R
\docType{data}
\name{pub_number}
\alias{pub_number}
\title{Data pub_number}
\format{An object of class \code{character} of length 10.}
\usage{
pub_number
}
\description{
A vector of ten epodoc format patent numbers for use in testing services such as ops_fulltext and ops_post.
}
\keyword{datasets}

