% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_pubs.R
\docType{data}
\name{pubs}
\alias{pubs}
\title{Data pubs}
\format{An object of class \code{character} of length 2.}
\usage{
pubs
}
\description{
Two WO patent document numbers (epodoc) for use in testing.
}
\keyword{datasets}

