% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_lens.R
\name{rename_lens}
\alias{rename_lens}
\title{Rename Patent Lens Columns}
\usage{
rename_lens(data)
}
\arguments{
\item{data}{a raw data.frame of patent results from the lens.}
}
\value{
a data frame with lower case names and underscores in place of
  spaces.
}
\description{
Convert Lens database column names from sentence case to lower
  case, replace spaces with underscores, shorten names with brackets and
  shorten classification names.
}
\details{
Uses tolower from base to convert all names to lower and then uses
dplyr rename_ to rename each column
}
\examples{
\dontrun{rename_lens(data)}
}

