% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_three_urls.R
\docType{data}
\name{three_urls}
\alias{three_urls}
\title{Data three_urls}
\format{An object of class \code{character} of length 3.}
\usage{
three_urls
}
\description{
A file containing a vector with three urls for a published data search (numbers service) for use with ops_iterate().
}
\keyword{datasets}

