% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_urls_1990_2000.R
\docType{data}
\name{urls_1990_2000}
\alias{urls_1990_2000}
\title{Data urls_1990_2000}
\format{An object of class \code{character} of length 9.}
\source{
EPO Open Patent Services through ops_urls ((urls_1990_2000 <-
  ops_urls("pizza", type = "ta", start = 1990, end = 2000))
}
\usage{
urls_1990_2000
}
\description{
A set of urls created using ops_urls() for a search of OPS biblios for the
term pizza between 1990-2000 for use in ops_iterate()
}
\keyword{datasets}

